CRUD up the sections

+++ routes/api/projectRoutes.js
@@ -4,11 +4,13 @@ const auth = require('../auth');
 
 const Project = mongoose.model('Project');
 
+router.use('/:projectId/sections', require('./sectionRoutes'));
+
 // POST create Project
 router.post('/', auth.required, async (req, res) => {
-  const { payload: { _id } } = req;
+  const userId = req.payload._id;
   const finalProject = new Project({
-    userId: _id,
+    userId,
     children: [
       {
         name: 'Getting Started',
@@ -30,11 +32,26 @@ router.post('/', auth.required, async (req, res) => {
   }
 });
 
-// GET Projects
-router.get('/', auth.required, async (req, res) => {
-  const { payload: { _id } } = req;
+// Get Project
+router.get('/:projectId', auth.optional, async (req, res) => {
+  const { params: { projectId } } = req;
   try {
-    const projects = await Project.find({ userId: _id });
+    const project = await Project.findOne({ _id: projectId });
+    res.send(project.toJSON());
+  } catch (err) {
+    res.json({
+      errors: {
+        project: 'Something went wrong',
+      },
+    });
+  }
+});
+
+// GET Projects index
+router.get('/', auth.required, async (req, res) => {
+  const userId = req.payload._id;
+  try {
+    const projects = await Project.find({ userId });
     res.send(projects);
   } catch (err) {
     res.json({
@@ -46,34 +63,16 @@ router.get('/', auth.required, async (req, res) => {
 });
 
 // DELETE remove Project
-router.delete('/:_id', async (req, res) => {
+router.delete('/:_id', auth.required, async (req, res) => {
+  const userId = req.payload._id;
   const { params: { _id } } = req;
   try {
-    const project = await Project.findOneAndDelete({ _id });
+    const project = await Project.findOneAndDelete({ userId, _id });
     res.send(project);
   } catch (err) {
     res.json({
       errors: {
         project: 'Something went wrong',
-      },
-    });
-  }
-});
-
-// POST add Section
-router.post('/:_id', async (req, res) => {
-  const { params: { _id } } = req;
-  const { body: { section } } = req;
-  try {
-    const finalProject = await Project.findOneAndUpdate(
-      { _id },
-      { $push: { children: section } },
-    );
-    res.send(finalProject);
-  } catch (err) {
-    res.json({
-      errors: {
-        section: 'Something went wrong',
       },
     });
   }
